# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  codeberg_org_azzet_azzetbe_internal_document.DocumentListResponse:
    properties:
      documents:
        items:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_document.DocumentResponse'
        type: array
      total:
        type: integer
    type: object
  codeberg_org_azzet_azzetbe_internal_document.DocumentResponse:
    properties:
      created_at:
        type: string
      document_type:
        type: string
      extracted_data:
        additionalProperties: true
        type: object
      extraction_confidence:
        type: number
      extraction_error:
        type: string
      extraction_status:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: string
      mime_type:
        type: string
      processed_at:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
      upload_status:
        type: string
      uploaded_at:
        type: string
      verification_status:
        type: string
      view_url:
        type: string
      workspace_id:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_document.PresignedUploadResponse:
    properties:
      document_id:
        type: string
      expires_in:
        type: integer
      file_key:
        type: string
      upload_url:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_document.UploadRequest:
    properties:
      document_type:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      mime_type:
        type: string
    type: object
...
  codeberg_org_azzet_azzetbe_internal_shared.APIResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise-grade accounting, tax, and finance platform. User routes
    use Bearer JWT from /auth/login. Workspace-scoped routes require X-Workspace-ID
    header. Admin routes under /admin use admin JWT from /admin/auth/login.
  title: Azzet API
  version: "1.0"
paths:
...
  /documents:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_document.DocumentListResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List workspace documents
      tags:
      - Document
    post:
      consumes:
      - application/json
      description: Get a presigned URL to upload a receipt/invoice for OCR processing
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Upload request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_document.UploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_document.PresignedUploadResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Request document upload URL
      tags:
      - Document
  /documents/{id}:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_document.DocumentResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get document detail
      tags:
      - Document
  /documents/{id}/confirm:
    post:
      description: Confirm that the file was uploaded to R2 and trigger OCR processing
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_document.DocumentResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Confirm document upload
      tags:
      - Document
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"